Let music21's own plumbing through, and give a style back that carries a page
An accidental, a beam, a duration and a lyric are not things a stream
holds, so they were replaced outright rather than installed — which meant
they inherited nothing at all, and every member music21 gives them
through `ProtoM21Object`, `StyleMixin` and the slots mixins simply went
missing: no `editorial`, no `id`, no `_getSlotsRecursive`. They are now
installed like everything else, over music21's own machinery classes.
`StyleMixin` cannot be a base of a pyo3 class — its `__slots__` will not
share a layout — so its editorial half is copied onto the installed class
instead, and `class_to_install` records what it built so the doctest
harness and `install_into_music21` agree on what a facade should build.

**`style` is music21's object now, not ours.** The facade used to hand
back a `Style` of its own that carried the colour and answered `None` to
every question about the page. music21's own notation code asks those
questions — `n.style.stemStyle`, `a.style.relativeX` — and a page it
cannot write on is a page it writes nothing to. So `note.style` is a real
`music21.style.NoteStyle`, `pitch.Accidental.style` a real `TextStyle`,
and the colour, which the crate does model, is written into that object
and read back out of it; `synced` folds it into the value, so an edit
through `n.style.color` is an edit to the note. The `Style` facade class
is gone.

Also found by the suite:

- `Pitch.fullName` names a written natural. music21 asks whether the pitch
  carries an accidental object, not whether it alters anything, so an
  explicit `Cn` is "C-natural in octave 4" — which is what the MEI reader
  builds and what its tests assert.
- A note answers as a quarter when nobody has timed it. The crate lets a
  note carry no duration; music21's `GeneralNote` always has one.
- `pitchFromDegree` counts round the scale. music21 reads every degree
  within the one octave its network holds, so the eighth degree is the
  tonic again rather than the octave above it, and the zeroth and the
  negative degrees count back from the top. The two hand-rolled folds
  that worked around the old behaviour are gone with it.
- `GenericInterval.complement` built an `Interval`, which only showed
  once the classes were installed and the cast had something to fail on.
- `Chord._notes`, `Volume`'s client surviving a copy, and a copy of a
  note, chord, accidental, beam or lyric carrying the style it was drawn
  with.
- A note and a chord freeze their ornaments and marks beside their value:
  music21 pickles every score it parses, and a thawed note with no
  articulations had lost what the score said — which is what made the
  fret indications vanish.
- `note.lyrics` is the list itself, since music21's MusicXML reader
  appends each verse to what it hands back.

music21's own suite, run against `install_into_music21()`, is down from
254 tests failing only under the crate to 225.

+++ python/src/scale.rs
@@ -850,7 +850,7 @@ impl ConcreteScale {
     ))]
     fn pitchFromDegree(
         &self,
-        degree: usize,
+        degree: i32,
         minPitch: Option<&Bound<'_, PyAny>>,
         maxPitch: Option<&Bound<'_, PyAny>>,
         direction: Option<&Bound<'_, PyAny>>,
